v5.3.7 add metadata

+++ main.py
@@ -26,6 +26,7 @@ parser.add_argument('--epochs', type=int, default=300)
 parser.add_argument('--weight_decay', type=float, default=0.0)
 parser.add_argument('--lr', type=float, default=0.0001)
 parser.add_argument('--grad_accum', type=int, default=1)
+parser.add_argument('--rmtree', type=bool, default=True)
 args = parser.parse_args()
 
 for task in mb.tasks:
@@ -106,7 +107,8 @@ for task in mb.tasks:
         task.record(fold, predictions)
         if torch.cuda.is_available():
             torch.cuda.empty_cache()
-    shutil.rmtree(name)
+    if args.rmtree:
+        shutil.rmtree(name)
 
 hyperparam = vars(args)
 hyperparam['grad_accum'] = grad_accum
